Revert "CODE MERGED"
This reverts commit be53be645f3bd8de4fa91777e8c254c80674e94e, reversing
changes made to d7a8050157d1a5d04a8684688098bbe0b9f501f7.

+++ lib/features/dashboard/data/model/user_dashboard.model.dart
@@ -249,7 +249,7 @@ class Table5 {
   final int leaveId;
   final DateTime startDate;
   final DateTime endDate;
-  final double noOfDays;
+  final int noOfDays;
   final String leaveTypeName;
   final String reason;
   final String message;
@@ -268,7 +268,7 @@ class Table5 {
     leaveId: parseValue<int>(json, "LeaveId"),
     startDate: parseValue<DateTime>(json, "StartDate"),
     endDate: parseValue<DateTime>(json, "EndDate"),
-    noOfDays: parseValue<double>(json, "NoOfDays"),
+    noOfDays: parseValue<int>(json, "NoOfDays"),
     leaveTypeName: parseValue<String>(json, "LeaveTypeName"),
     reason: parseValue<String>(json, "Reason"),
     message: parseValue<String>(json, "Message"),
